use Git::PushInitial to work with all push configs
Use my new ::Plugin::Git::PushInitial to always push the initial repo
after minting.

This makes it so that the user's push.default config no longer matters,
if it's 'tracking' there will be a remote branch to match because the
initial repo was pushed.

Remove the ugly push examples from SYNOPSIS (that was the point.)